+++ libs/cardano-ledger-api/test/Test/Cardano/Ledger/Api/State/Imp/QuerySpec.hs
@@ -10,11 +10,11 @@
 module Test.Cardano.Ledger.Api.State.Imp.QuerySpec where
 
 import Cardano.Ledger.Api.State.Query (
-  CommitteeMemberState (..),
-  CommitteeMembersState (..),
   HotCredAuthStatus (..),
   MemberStatus (..),
   NextEpochChange (..),
+  QueryResultCommitteeMemberState (..),
+  QueryResultCommitteeMembersState (..),
   queryCommitteeMembersState,
   queryDRepDelegations,
   queryDRepState,
@@ -172,7 +172,7 @@ spec = withEachEraVersion @era $ do
           UpdateCommittee SNothing mempty (Map.singleton c1 (EpochNo 4321)) (1 %! 1)
         hk1 <- registerCommitteeHotKey c1
         expectQueryResult (Set.singleton c1) mempty mempty $
-          [(c1, CommitteeMemberState (MemberAuthorized hk1) Unrecognized Nothing ToBeRemoved)]
+          [(c1, QueryResultCommitteeMemberState (MemberAuthorized hk1) Unrecognized Nothing ToBeRemoved)]
         passEpoch
         expectQueryResult (Set.singleton c1) mempty mempty Map.empty
 
@@ -196,10 +196,11 @@ spec = withEachEraVersion @era $ do
         passEpoch
         hk1 <- registerCommitteeHotKey c1
         expectQueryResult (Set.singleton c1) mempty mempty $
-          [(c1, CommitteeMemberState (MemberAuthorized hk1) Unrecognized Nothing ToBeEnacted)]
+          [(c1, QueryResultCommitteeMemberState (MemberAuthorized hk1) Unrecognized Nothing ToBeEnacted)]
         passEpoch
         expectQueryResult (Set.singleton c1) mempty mempty $
-          [(c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)]
+          [ (c1, QueryResultCommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)
+          ]
 
   it "queryDRepDelegationState" $ do
     (credDrep, delegator, _) <- setupSingleDRep 1_000_000
@@ -272,26 +273,26 @@ spec = withEachEraVersion @era $ do
     expectMembers $ Map.keysSet newMembers
     -- members for which the expiration epoch is the current epoch are `ToBeExpired`
     expectNoFilterQueryResult
-      [ (c1, CommitteeMemberState MemberNotAuthorized Active (Just c1Expiry) NoChangeExpected)
-      , (c2, CommitteeMemberState MemberNotAuthorized Active (Just c2Expiry) ToBeExpired)
-      , (c3, CommitteeMemberState MemberNotAuthorized Active (Just c3Expiry) NoChangeExpected)
-      , (c4, CommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) NoChangeExpected)
+      [ (c1, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c1Expiry) NoChangeExpected)
+      , (c2, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c2Expiry) ToBeExpired)
+      , (c3, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c3Expiry) NoChangeExpected)
+      , (c4, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) NoChangeExpected)
       ]
     -- hot cred status of members with registered hot keys becomes `MemberAuthorized`
     hk1 <- registerCommitteeHotKey c1
     hk2 <- registerCommitteeHotKey c2
     expectNoFilterQueryResult
-      [ (c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)
-      , (c2, CommitteeMemberState (MemberAuthorized hk2) Active (Just c2Expiry) ToBeExpired)
-      , (c3, CommitteeMemberState MemberNotAuthorized Active (Just c3Expiry) NoChangeExpected)
-      , (c4, CommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) NoChangeExpected)
+      [ (c1, QueryResultCommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)
+      , (c2, QueryResultCommitteeMemberState (MemberAuthorized hk2) Active (Just c2Expiry) ToBeExpired)
+      , (c3, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c3Expiry) NoChangeExpected)
+      , (c4, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) NoChangeExpected)
       ]
     expectQueryResult
       [c2, c3, c5]
       mempty
       [Active, Unrecognized]
-      [ (c3, CommitteeMemberState MemberNotAuthorized Active (Just c3Expiry) NoChangeExpected)
-      , (c2, CommitteeMemberState (MemberAuthorized hk2) Active (Just c2Expiry) ToBeExpired)
+      [ (c3, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c3Expiry) NoChangeExpected)
+      , (c2, QueryResultCommitteeMemberState (MemberAuthorized hk2) Active (Just c2Expiry) ToBeExpired)
       ]
 
     c3Anchor <- arbitrary
@@ -309,11 +310,18 @@ spec = withEachEraVersion @era $ do
     -- registering a hot key for a credential that's not part of the committee will yield `Unrecognized` member status
     -- and expected change of `ToBeRemoved`
     expectNoFilterQueryResult
-      [ (c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)
-      , (c2, CommitteeMemberState (MemberAuthorized hk2) Active (Just c2Expiry) ToBeExpired)
-      , (c3, CommitteeMemberState (MemberResigned (Just c3Anchor)) Active (Just c3Expiry) NoChangeExpected)
-      , (c4, CommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) NoChangeExpected)
-      , (c5, CommitteeMemberState (MemberAuthorized hk5) Unrecognized Nothing ToBeRemoved)
+      [ (c1, QueryResultCommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)
+      , (c2, QueryResultCommitteeMemberState (MemberAuthorized hk2) Active (Just c2Expiry) ToBeExpired)
+      ,
+        ( c3
+        , QueryResultCommitteeMemberState
+            (MemberResigned (Just c3Anchor))
+            Active
+            (Just c3Expiry)
+            NoChangeExpected
+        )
+      , (c4, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) NoChangeExpected)
+      , (c5, QueryResultCommitteeMemberState (MemberAuthorized hk5) Unrecognized Nothing ToBeRemoved)
       ]
     expectQueryResult
       [c2, c3, c5]
@@ -321,17 +329,27 @@ spec = withEachEraVersion @era $ do
       [Unrecognized]
       ( Map.singleton
           c5
-          (CommitteeMemberState (MemberAuthorized hk5) Unrecognized Nothing ToBeRemoved)
+          (QueryResultCommitteeMemberState (MemberAuthorized hk5) Unrecognized Nothing ToBeRemoved)
       )
 
     passEpoch -- epoch 3
     -- the `Unrecognized` member gets removed from the query result
     -- the member which in the previous epoch was expected `ToBeEpired`, has now MemberStatus `Expired` and `NoChangeExpected`
     expectNoFilterQueryResult
-      [ (c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)
-      , (c2, CommitteeMemberState (MemberAuthorized hk2) Expired (Just c2Expiry) NoChangeExpected)
-      , (c3, CommitteeMemberState (MemberResigned (Just c3Anchor)) Active (Just c3Expiry) NoChangeExpected)
-      , (c4, CommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) NoChangeExpected)
+      [ (c1, QueryResultCommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) NoChangeExpected)
+      ,
+        ( c2
+        , QueryResultCommitteeMemberState (MemberAuthorized hk2) Expired (Just c2Expiry) NoChangeExpected
+        )
+      ,
+        ( c3
+        , QueryResultCommitteeMemberState
+            (MemberResigned (Just c3Anchor))
+            Active
+            (Just c3Expiry)
+            NoChangeExpected
+        )
+      , (c4, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) NoChangeExpected)
       ]
 
     -- elect new committee to be: c1 (term extended ), c3 (no changes), c4 (term shortened, expiring next epoch), c6, c7 (new)
@@ -361,20 +379,41 @@ spec = withEachEraVersion @era $ do
     -- members that are part of both current and next committee have `NoChangeExpected` or `TermAdjusted`
     -- members that part of only next committee are `ToBeEnacted`
     expectNoFilterQueryResult
-      [ (c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) (TermAdjusted c1NewExpiry))
-      , (c2, CommitteeMemberState (MemberAuthorized hk2) Expired (Just c2Expiry) ToBeRemoved)
-      , (c3, CommitteeMemberState (MemberResigned (Just c3Anchor)) Active (Just c3Expiry) NoChangeExpected)
+      [
+        ( c1
+        , QueryResultCommitteeMemberState
+            (MemberAuthorized hk1)
+            Active
+            (Just c1Expiry)
+            (TermAdjusted c1NewExpiry)
+        )
+      , (c2, QueryResultCommitteeMemberState (MemberAuthorized hk2) Expired (Just c2Expiry) ToBeRemoved)
+      ,
+        ( c3
+        , QueryResultCommitteeMemberState
+            (MemberResigned (Just c3Anchor))
+            Active
+            (Just c3Expiry)
+            NoChangeExpected
+        )
       , -- though its term was adjusted, `ToBeExpired` takes precedence
-        (c4, CommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) ToBeExpired)
-      , (c6, CommitteeMemberState (MemberAuthorized hk6) Unrecognized Nothing ToBeEnacted)
-      , (c7, CommitteeMemberState MemberNotAuthorized Unrecognized Nothing ToBeEnacted)
-      , (c8, CommitteeMemberState (MemberAuthorized hk8) Unrecognized Nothing ToBeRemoved)
+        (c4, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c4Expiry) ToBeExpired)
+      , (c6, QueryResultCommitteeMemberState (MemberAuthorized hk6) Unrecognized Nothing ToBeEnacted)
+      , (c7, QueryResultCommitteeMemberState MemberNotAuthorized Unrecognized Nothing ToBeEnacted)
+      , (c8, QueryResultCommitteeMemberState (MemberAuthorized hk8) Unrecognized Nothing ToBeRemoved)
       ]
     expectQueryResult
       [c1]
       mempty
       mempty
-      [ (c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1Expiry) (TermAdjusted c1NewExpiry))
+      [
+        ( c1
+        , QueryResultCommitteeMemberState
+            (MemberAuthorized hk1)
+            Active
+            (Just c1Expiry)
+            (TermAdjusted c1NewExpiry)
+        )
       ]
     expectQueryResult
       [c2]
@@ -382,18 +421,31 @@ spec = withEachEraVersion @era $ do
       [Expired]
       ( Map.singleton
           c2
-          (CommitteeMemberState (MemberAuthorized hk2) Expired (Just c2Expiry) ToBeRemoved)
+          (QueryResultCommitteeMemberState (MemberAuthorized hk2) Expired (Just c2Expiry) ToBeRemoved)
       )
 
     passNEpochs 2 -- epoch 6
     -- the new committee is in place with the adjusted terms
     expectMembers [c1, c3, c4, c6, c7]
     expectNoFilterQueryResult
-      [ (c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1NewExpiry) NoChangeExpected)
-      , (c3, CommitteeMemberState (MemberResigned (Just c3Anchor)) Active (Just c3Expiry) NoChangeExpected)
-      , (c4, CommitteeMemberState MemberNotAuthorized Expired (Just c4NewExpiry) NoChangeExpected)
-      , (c6, CommitteeMemberState (MemberAuthorized hk6) Active (Just c6Expiry) ToBeExpired)
-      , (c7, CommitteeMemberState MemberNotAuthorized Active (Just c7Expiry) NoChangeExpected)
+      [
+        ( c1
+        , QueryResultCommitteeMemberState (MemberAuthorized hk1) Active (Just c1NewExpiry) NoChangeExpected
+        )
+      ,
+        ( c3
+        , QueryResultCommitteeMemberState
+            (MemberResigned (Just c3Anchor))
+            Active
+            (Just c3Expiry)
+            NoChangeExpected
+        )
+      ,
+        ( c4
+        , QueryResultCommitteeMemberState MemberNotAuthorized Expired (Just c4NewExpiry) NoChangeExpected
+        )
+      , (c6, QueryResultCommitteeMemberState (MemberAuthorized hk6) Active (Just c6Expiry) ToBeExpired)
+      , (c7, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c7Expiry) NoChangeExpected)
       ]
     expectQueryResult
       Set.empty
@@ -422,10 +474,10 @@ spec = withEachEraVersion @era $ do
     passEpoch -- epoch 7
     -- members whose term changed have next epoch change `TermAdjusted`
     expectNoFilterQueryResult
-      [ (c1, CommitteeMemberState (MemberAuthorized hk1) Active (Just c1NewExpiry) ToBeRemoved)
+      [ (c1, QueryResultCommitteeMemberState (MemberAuthorized hk1) Active (Just c1NewExpiry) ToBeRemoved)
       ,
         ( c3
-        , CommitteeMemberState
+        , QueryResultCommitteeMemberState
             (MemberResigned (Just c3Anchor))
             Active
             (Just c3Expiry)
@@ -433,21 +485,42 @@ spec = withEachEraVersion @era $ do
         )
       ,
         ( c4
-        , CommitteeMemberState MemberNotAuthorized Expired (Just c4NewExpiry) (TermAdjusted c4NewNewExpiry)
+        , QueryResultCommitteeMemberState
+            MemberNotAuthorized
+            Expired
+            (Just c4NewExpiry)
+            (TermAdjusted c4NewNewExpiry)
         )
-      , (c6, CommitteeMemberState (MemberAuthorized hk6) Expired (Just c6Expiry) (TermAdjusted c6NewExpiry))
-      , (c7, CommitteeMemberState MemberNotAuthorized Active (Just c7Expiry) ToBeExpired)
+      ,
+        ( c6
+        , QueryResultCommitteeMemberState
+            (MemberAuthorized hk6)
+            Expired
+            (Just c6Expiry)
+            (TermAdjusted c6NewExpiry)
+        )
+      , (c7, QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c7Expiry) ToBeExpired)
       ]
     passEpoch -- epoch 8
     expectMembers [c3, c4, c6, c7]
     expectNoFilterQueryResult
       [
         ( c3
-        , CommitteeMemberState (MemberResigned (Just c3Anchor)) Active (Just c3NewExpiry) NoChangeExpected
+        , QueryResultCommitteeMemberState
+            (MemberResigned (Just c3Anchor))
+            Active
+            (Just c3NewExpiry)
+            NoChangeExpected
         )
-      , (c4, CommitteeMemberState MemberNotAuthorized Active (Just c4NewNewExpiry) NoChangeExpected)
-      , (c6, CommitteeMemberState (MemberAuthorized hk6) Active (Just c6NewExpiry) NoChangeExpected)
-      , (c7, CommitteeMemberState MemberNotAuthorized Expired (Just c7Expiry) NoChangeExpected)
+      ,
+        ( c4
+        , QueryResultCommitteeMemberState MemberNotAuthorized Active (Just c4NewNewExpiry) NoChangeExpected
+        )
+      ,
+        ( c6
+        , QueryResultCommitteeMemberState (MemberAuthorized hk6) Active (Just c6NewExpiry) NoChangeExpected
+        )
+      , (c7, QueryResultCommitteeMemberState MemberNotAuthorized Expired (Just c7Expiry) NoChangeExpected)
       ]
   where
     expectQueryResult ::
@@ -455,22 +528,22 @@ spec = withEachEraVersion @era $ do
       Set.Set (Credential ColdCommitteeRole) ->
       Set.Set (Credential HotCommitteeRole) ->
       Set.Set MemberStatus ->
-      Map.Map (Credential ColdCommitteeRole) CommitteeMemberState ->
+      Map.Map (Credential ColdCommitteeRole) QueryResultCommitteeMemberState ->
       ImpTestM era ()
     expectQueryResult ckFilter hkFilter statusFilter expResult = do
       nes <- use impNESL
-      let CommitteeMembersState {csCommittee} =
+      let QueryResultCommitteeMembersState {qrcmsCommittee} =
             queryCommitteeMembersState
               ckFilter
               hkFilter
               statusFilter
               nes
       impAnn "Expecting query result" $
-        csCommittee `shouldBe` expResult
+        qrcmsCommittee `shouldBe` expResult
 
     expectNoFilterQueryResult ::
       HasCallStack =>
-      Map.Map (Credential ColdCommitteeRole) CommitteeMemberState ->
+      Map.Map (Credential ColdCommitteeRole) QueryResultCommitteeMemberState ->
       ImpTestM era ()
     expectNoFilterQueryResult =
       expectQueryResult mempty mempty mempty